MOD: Modify compatibility of kafka plugin and expand operationName
